# make sim_tree faster for large N